+++ src/Compiler/IL/LIRToILCompiler.TempsLocals.cs
@@ -1967,40 +1967,27 @@ internal sealed partial class LIRToILCompiler
                     return true;
                 }
 
-                foreach (var arg in callRuntimeServices.Arguments)
+                if (callRuntimeServices.ParameterTypes is { Count: var parameterTypeCount }
+                    && parameterTypeCount != callRuntimeServices.Arguments.Count)
                 {
-                    EmitLoadTemp(arg, ilEncoder, allocation, methodDescriptor);
+                    throw new InvalidOperationException(
+                        $"RuntimeServices call '{callRuntimeServices.MethodName}' has {parameterTypeCount} parameter types for {callRuntimeServices.Arguments.Count} arguments.");
+                }
 
-                    if (GetTempStorage(arg) is { } storage
-                        && storage.Kind == ValueStorageKind.UnboxedValue
-                        && storage.ClrType != null)
-                    {
-                        ilEncoder.OpCode(ILOpCode.Box);
-                        if (storage.ClrType == typeof(double))
-                        {
-                            ilEncoder.Token(_bclReferences.DoubleType);
-                        }
-                        else if (storage.ClrType == typeof(bool))
-                        {
-                            ilEncoder.Token(_bclReferences.BooleanType);
-                        }
-                        else if (storage.ClrType == typeof(int))
-                        {
-                            ilEncoder.Token(_bclReferences.Int32Type);
-                        }
-                        else
-                        {
-                            throw new NotSupportedException($"Unsupported unboxed type for RuntimeServices call: {storage.ClrType}");
-                        }
-                    }
+                var paramTypes = new Type[callRuntimeServices.Arguments.Count];
+                for (int i = 0; i < callRuntimeServices.Arguments.Count; i++)
+                {
+                    var parameterType = callRuntimeServices.ParameterTypes?[i] ?? typeof(object);
+                    paramTypes[i] = parameterType;
+                    EmitLoadTempForParameter(
+                        callRuntimeServices.Arguments[i],
+                        parameterType,
+                        ilEncoder,
+                        allocation,
+                        methodDescriptor);
                 }
 
                 ilEncoder.OpCode(ILOpCode.Call);
-                var paramTypes = new Type[callRuntimeServices.Arguments.Count];
-                for (int i = 0; i < paramTypes.Length; i++)
-                {
-                    paramTypes[i] = typeof(object);
-                }
                 var methodRef = _memberRefRegistry.GetOrAddMethod(
                     typeof(JavaScriptRuntime.RuntimeServices),
                     callRuntimeServices.MethodName,